+++ src/Karteikartensystem/1. Code/Model/Stapel/Stapel.cs
@@ -4,7 +4,7 @@ namespace Model
 {
     public class Stapel
     {
-        public int Id { get; set; } // in long ändern
+        public long Id { get; set; } // in long ändern
         public string Name { get; set; }
         public List<Karte> Karten { get; set; }
         public long GelernteZeitInMinuten { get; set; }
